Reenable emacs
Decided I want to give emacs another shot. Going back to use-package because it seems easier
and I know it works with Hasklig (not sure how to get hasklig working with just require).
Additionally, we can remove the ':ensure t' logic to prevent use-package from pulling in
missing deps (since we want to use Nix as much as possible).