Remove dunst
Having dunst as an option lead to surprising behavior when it
interferred with deadd. It _appears_ that notify-send (which we
use for testing) was launching dunst when no other server was
present.

Moreover, for reasons that are still unclear, it was blocking deadd
on startup, even though we were not inclding the dunst.nix file in
our config. We do not need it currently, so it can be safely removed.