Fix polybar
Polybar has been flaky with the desktop workspaces and ewmh. Adding the
asynchronous delay appears to solve this (i.e. we have a race condition),
though polybar does seem to spawn at an irregular time now).

Possibly the asynchronous spawning itself is enough to fix our problem,
but this is unclear at this time